requirements.txt
fastapi==0.92.0
ujson==5.7.0
uvicorn==0.20.0
SQLAlchemy==2.0.4
SQLAlchemy-Utils==0.40.0
alembic==1.9.4
ipython==7.29.0
environs==9.5.0
ipdb==0.13.9
asyncpg==0.27.0
google-cloud-texttospeech==2.14.1
google-cloud-speech==2.17.3
gcloud-aio-storage==8.0.0
click==8.1.3
python-multipart==0.0.5
pre-commit==3.1.0
